Update README.md
A loop for user entry of dates and display of newly formatted dates.  
User will be prompted to enter a date. The following is an example of the format that each date must follow when entered: March 31, 2021. 
Each correctly entered date will be reformatted. Output each correct date as: m/d/yyyy. (e.g. 3/31/2021) 
Any date not entered in the correct format will display a message. 
The loop ends when a -1 is entered on a line alone.